requirements.txt
fastapi
uvicorn[standard]
streamlit
python-multipart
pydantic-settings
sqlalchemy
psycopg2-binary
requests

# Vector DB
faiss-cpu
chromadb
sentence-transformers

# OCR
paddleocr
paddlepaddle

# Parsing
pymupdf
pdfplumber

# NLP
spacy

# LLM (Direct)
ollama

# Utils
python-dotenv
langfuse
structlog
cryptography
passlib[bcrypt]
python-jose[cryptography]

# ML
scikit-learn
joblib
numpy
